Refactoring: GridCreatingConditionTriangle::DefinePolyLineNewPointCommand implementation is separated from gridcreatingconditiontriangle.cpp.
